requirements.txt
Pillow~=7.0.0
numpy~=1.22.4
matplotlib~=3.5.2